Add openFixture method to OpensDocumentsTrait
