+++ tests/Handler/OpensDocumentsTrait.php
@@ -29,4 +29,16 @@ trait OpensDocumentsTrait
             ],
         ]));
     }
+
+    private function openFixture(string $fixturePath): string
+    {
+        $fullPath = dirname(__DIR__) . '/Fixtures/' . $fixturePath;
+        $content = file_get_contents($fullPath);
+        assert($content !== false, "Fixture not found: $fixturePath");
+
+        $uri = 'file:///fixtures/' . $fixturePath;
+        $this->openDocument($uri, $content);
+
+        return $uri;
+    }
 }
